+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/CHANGELOG.md
@@ -10,6 +10,9 @@
   telemetry processing failures do not escape request stop processing.
   ([#4344](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4344))
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
 ## 1.15.2
 
 Released 2026-Apr-21