+++ src/OpenTelemetry.Instrumentation.Hangfire/CHANGELOG.md
@@ -6,6 +6,9 @@
   the previous ambient baggage is restored after job completion.
   ([#4288](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4288))
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
 ## 1.15.1-beta.1
 
 Released 2026-Apr-21